requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
APScheduler==3.9.1.post1
requests==2.28.1
python-dotenv==0.21.0
pdfkit==0.6.1
jinja2==3.1.2
google-api-python-client==2.92.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
bcrypt==4.0.1
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.5
SQLAlchemy==2.0.3
numpy==1.21.6
pandas==1.3.5
PyJWT==2.6.0
